Move method and kernel implementation in separate files
