+++ pandas/core/window/rolling.py
@@ -39,7 +39,7 @@ from pandas._typing import Axis, FrameOrSeries, Scalar
 from pandas.core.base import DataError, PandasObject, SelectionMixin
 import pandas.core.common as com
 from pandas.core.index import Index, ensure_index
-from pandas.core.window.aggregators import rolling_mean_method, rolling_mean_class
+from pandas.core.window.aggregators.methods import rolling_mean_method
 from pandas.core.window.common import (
     _check_min_periods,
     _doc_template,
